--- docs/.vscode/extensions.json
@@ -1,7 +1,0 @@
-{
-  "recommendations": [
-    "DavidAnson.vscode-markdownlint",
-    "EditorConfig.EditorConfig",
-    "streetsidesoftware.code-spell-checker"
-  ]
-}
