+++ internal/warpc/genavif/go.mod
@@ -1,0 +1,3 @@
+module gohugoio/hugo/internal/warpc/genavif
+
+go 1.21
