+++ resources/testdata/bep/license.txt
@@ -1,0 +1,1 @@
+All images in this folder are copyrighted Bjørn Erik Pedersen (2026), Creative Commons Attribution-Share Alike 4.0 International license.
